SDL Fixes (#571)
* SDL Fixes

11065424 - Defect : InvalidParamValue1, Component : c-utility\adapters\tlsio_schannel.c
11065425 - Defect : Unknown, Component : c-utility\src\crt_abstractions.c

* Update tlsio_schannel.c

removed extra line

+++ src/crt_abstractions.c
@@ -655,7 +655,7 @@ long double strtold_s(const char* nptr, char** endptr)
             break;
         case FST_NAN:
             /*Codes_SRS_CRT_ABSTRACTIONS_21_034: [If the string is 'NAN' or 'NAN(...)' (ignoring case), the strtold_s must return 0.0 and points endptr to the first character after the 'NAN' sequence.]*/
-            result = (long double)NAN;
+            result = (long double)((float)NAN);
             break;
         case FST_NUMBER:
             if ((exponential != DBL_MAX_10_EXP || (fraction <= 1.7976931348623158)) &&
